+++ Lib/CAT_Surf.c
@@ -1265,7 +1265,7 @@ check_polygons_shape_integrity(polygons_struct *polygons, Point new_points[])
     int vertidx, ptidx, poly, size;
     int n_nb, neighbours[MAX_NEIGHBOURS];
     int n_errors, n_bad_points;
-    float base_length, curv_factor;
+    double base_length, curv_factor;
     Vector normal;
     BOOLEAN interior_flag;
